event_spec: world-ify prequel
modularize genstubs convenience function

Also drops the unused reference to match (can always bring it back if we
end up actually using it)
