+++ spec/event/event_spec.lua
@@ -2,7 +2,7 @@ require('spec/setup/defines')
 require('stdlib/utils/string')
 require('stdlib/utils/table')
 
-local match = require('luassert.match')
+local genstubs = require('spec/utils/stub_factory')
 
 local Event
 
@@ -19,21 +19,6 @@ local function fire(e)
     if factorio_event_registry[e.name] then
         return factorio_event_registry[e.name](e)
     end
-end
-
--- convenience function to generate a pile of unique stubs, i.e.:
--- a,b,c = genstubs(3)
--- a('foo')
--- assert.stub(a).was.called_with('foo')
--- assert.stub(b).was.not.called()
-local function genstubs(count)
-    count = not count and 1 or count
-    local tmocks = {}
-    while count > 0 do
-        table.insert(tmocks, function() end)
-        count = count - 1
-    end
-    return table.unpack(mock(tmocks, true))
 end
 
 describe("Event",
